Remove BMA study fetch when starting poll
We have thus far made a decision that when the SA is unable to access
the BMA, it should fail loudly so we get a clear signal in deployments.
Unfortunately, in practice it seems that we may lose BMA access in
environments for various reasons. This change prevents the SA from
cycling rapidly in failures so we can maintain "fail loudly" while
avoiding high CPU and costs associated with immediately failing on
launch.